fix: ajuste em evento de click do redirect

+++ src/app/pages/home-container/home-container.component.ts
@@ -108,6 +108,7 @@ export class HomeContainerComponent{
   }
 
   redirectToNewConversation(){
+    if(!this.authService.isLoggedIn())return;
     sessionStorage.clear();
     this.selectedConversationId = '';
     this.router.navigate(['/conversation']);
